Refactor EntryImage
