+++ src/components/EntryImage/EntryImage.js
@@ -1,21 +1,25 @@
-import React, { useState} from 'react'
+import React, { useState } from 'react'
+import { Redirect } from 'react-router-dom';
+import getImages from '../../apiCalls';
 import './EntryImage.css'
-import ImageSearch from '../ImageSearch/ImageSearch'
 
 const EntryImage = ({ currentEntryId, journalEntries, setJournalEntries }) => {
   const [searchValue, setSearchValue] = useState("")
   const [selectedImage, setSelectedImage] = useState({})
   const [selectedImageIndex, setSelectedImageIndex] = useState(-1)
+  const [searchResults, setSearchResults] = useState([])
+  const [formSubmitted, setFormSubmitted] = useState(false)
+  const [error, setError] = useState(null)
 
   const addImageToJournalEntry = () => {
     const updatedJournalEntries = journalEntries.map((entry) => {
       if(entry.id === currentEntryId){
-        console.log(selectedImage)
         const updatedEntry = {...entry}
         updatedEntry.image.default = false
         updatedEntry.image.urls.regular = selectedImage.urls.regular;
         updatedEntry.image.urls.small= selectedImage.urls.small;
         updatedEntry.image.urls.thumb= selectedImage.urls.thumb;
+        updatedEntry.image.urls.raw= selectedImage.urls.raw;
         updatedEntry.image.altText=selectedImage.alt_description;
 
         return updatedEntry
@@ -24,6 +28,52 @@ const EntryImage = ({ currentEntryId, journalEntries, setJournalEntries }) => {
     })
     setJournalEntries(updatedJournalEntries)
   }
+
+  const getSearchImages = (event) => {
+    event.preventDefault()
+    getImages(searchValue)
+      .then(data => {
+        setSearchResults(data.results)
+        setError(null)
+      })
+      .catch(err => {
+        setError('Failed to fetch images.');
+        console.log(err)
+      })
+  }
+
+  const selectImage = (image, index) => {
+    setSelectedImage(image)
+    setSelectedImageIndex(index)
+  }
+
+  const handleSave = (event) => {
+    event.preventDefault()
+    addImageToJournalEntry()
+    setSearchValue("")
+    setSearchResults([])
+    setSelectedImage({})
+    setSelectedImageIndex(-1)
+    setFormSubmitted(true)
+  }
+
+  const resultImages = searchResults.map((result, index) => {
+    return (
+      <label key={index + 1}>
+        <input
+          type="radio"
+          name="image-one"
+          value={result}
+          checked={selectedImageIndex === index}
+          onChange={() => selectImage(result, index)} />
+        <img
+          className="result-image"
+          key={result.id}
+          src={result.urls.raw + '&h=550&dpr=2'}
+          alt={result.alt_description} />
+      </label>
+    )
+  })
 
   return (
     <div className="img-entry-wrapper">
@@ -39,13 +89,13 @@ const EntryImage = ({ currentEntryId, journalEntries, setJournalEntries }) => {
             value={searchValue}
             onChange={(event) => setSearchValue(event.target.value)} />
         </div>
-        <ImageSearch
-          searchValue={searchValue}
-          setSearchValue={setSearchValue}
-          setSelectedImage={setSelectedImage}
-          selectedImageIndex={selectedImageIndex}
-          setSelectedImageIndex={setSelectedImageIndex}
-          addImageToJournalEntry={addImageToJournalEntry} />
+        <button onClick={(event)=>getSearchImages(event)} disabled={!searchValue}>Search</button>
+        <div>
+          {formSubmitted && <Redirect to="/journal" />}
+          {error && <p>{error}</p>}
+          {resultImages}
+          <button onClick={(event) => { handleSave(event) }}>Save</button>
+        </div>
       </form>
     </div>
   )
