--- src/components/ImageSearch/ImageSearch.css
@@ -1,3 +1,0 @@
-.result-image{
-  height: 100px;
-}
